Move fuzz tests to standalone test file
