+++ sdks/go/pkg/beam/core/graph/coder/int_test.go
@@ -86,43 +86,3 @@ func TestEncodeDecodeInt32(t *testing.T) {
 		}
 	}
 }
-
-func FuzzEncodeDecodeUInt64(f *testing.F) {
-	f.Add(uint64(42))
-	f.Fuzz(func(t *testing.T, b uint64) {
-		var buf bytes.Buffer
-		err := EncodeUint64(b, &buf)
-		if err != nil {
-			return
-		}
-
-		got, err := DecodeUint64(&buf)
-		if err != nil {
-			t.Fatalf("failed to decode bytes, got %v", err)
-		}
-
-		if got != b {
-			t.Errorf("decoded output does not match input: got %v, want %v", got, b)
-		}
-	})
-}
-
-func FuzzEncodeDecodeInt32(f *testing.F) {
-	f.Add(int32(42))
-	f.Fuzz(func(t *testing.T, b int32) {
-		var buf bytes.Buffer
-		err := EncodeInt32(b, &buf)
-		if err != nil {
-			return
-		}
-
-		got, err := DecodeInt32(&buf)
-		if err != nil {
-			t.Fatalf("failed to decode bytes, got %v", err)
-		}
-
-		if got != b {
-			t.Errorf("decoded output does not match input: got %v, want %v", got, b)
-		}
-	})
-}
